requirements.txt
certifi==2021.10.8
joblib==1.1.0
scikit-learn==1.0.2
scipy==1.7.3
threadpoolctl==3.1.0
wincertstore==0.2
#xgboost==1.5.2
xgboost
streamlit==1.5.0
#numpy==1.20.2
pandas
plotly==5.3.1
nltk==3.6.7
#tensorflow==2.8.0
tensorflow
tflearn==0.5.0

beautifulsoup4==4.9.3
newspaper3k==0.2.8
pygooglenews==0.1.2
requests==2.26.0
#streamlit==0.85.1
numpy==1.19.4
# the following is for fixing Twint, see https://github.com/twintproject/twint/issues/915
git+https://github.com/twintproject/twint.git@origin/master#egg=twint
pickle5==0.0.11
#nltk==3.6.2
